⬆️ upgrade example to latest and api to ffmpeg 6

+++ lib/src/export/ffmpeg_export_config.dart
@@ -114,7 +114,7 @@ abstract class FFmpegVideoEditorConfig {
   /// });
   /// ```
   /// Returns the [double] progress value between 0.0 and 1.0.
-  double getFFmpegProgress(int time) {
+  double getFFmpegProgress(double time) {
     final double progressValue =
         time / controller.trimmedDuration.inMilliseconds;
     return progressValue.clamp(0.0, 1.0);
